Mockup for UI

+++ Source/oled.h
@@ -6,7 +6,7 @@
 #endif
 
   void oled_init(SPI_HandleTypeDef *spi);
-  void oled_write(const uint8_t* pixels, uint16_t size);
+  void oled_write(const uint8_t* pixels, uint32_t size);
 
 #ifdef __cplusplus
 }
